refactor: rename demo and config classes for improved package structure and clarity
