+++ src/main/java/org/thingai/app/Main.java
@@ -1,5 +1,7 @@
 package org.thingai.app;
 
+import eventimpl.fanroc.FanrocRankingStrategy;
+import eventimpl.fanroc.FanrocScore;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.servlet.context.ServletWebServerApplicationContext;
@@ -10,7 +12,6 @@ import org.springframework.core.io.support.ResourcePropertySource;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.thingai.app.scoringservice.ScoringService;
 import org.thingai.base.log.ILog;
-import org.thingai.app.fanroc.*;
 
 import java.net.InetAddress;
 import java.net.UnknownHostException;
